+++ oneflow/core/operator/conv_op.h
@@ -50,8 +50,8 @@ class ConvOp : public Operator {
   int32_t MaxModelSplitNum() const override;
 
  private:
-  void InferOutBlobModelSplitAxis(std::function<int64_t*(const std::string&)> ModelSplitAxis4BnInOp,
-                                  std::function<int64_t(const std::string&)> ShapeNumAxes4BnInOp,
+  void InferOutBlobModelSplitAxis(std::function<int32_t*(const std::string&)> ModelSplitAxis4BnInOp,
+                                  std::function<int32_t(const std::string&)> ShapeNumAxes4BnInOp,
                                   const ParallelContext* parallel_context) const override {
     NaiveInferOutBlobModelSplitAxis(ModelSplitAxis4BnInOp, ShapeNumAxes4BnInOp, parallel_context);
   }
