+++ oneflow/core/operator/reduce_gather_op.h
@@ -18,8 +18,8 @@ class ReduceGatherOp final : public Operator {
                       const ParallelContext* parallel_ctx) const override;
 
  private:
-  void InferOutBlobModelSplitAxis(std::function<int64_t*(const std::string&)> ModelSplitAxis4BnInOp,
-                                  std::function<int64_t(const std::string&)> ShapeNumAxes4BnInOp,
+  void InferOutBlobModelSplitAxis(std::function<int32_t*(const std::string&)> ModelSplitAxis4BnInOp,
+                                  std::function<int32_t(const std::string&)> ShapeNumAxes4BnInOp,
                                   const ParallelContext* parallel_context) const override {
     UNIMPLEMENTED();
   }
